Unnecessary depencencies, clippy, cosmetics

+++ src/ui/pages/gpu.rs
@@ -387,7 +387,7 @@ impl ResGPU {
         imp.temperature.graph().set_visible(temperature.is_some());
 
         if let Some(temperature) = temperature {
-            let temperature_string = convert_temperature(*temperature as f64);
+            let temperature_string = convert_temperature(*temperature);
 
             let highest_temperature_string =
                 convert_temperature(imp.temperature.graph().get_highest_value());
@@ -398,7 +398,7 @@ impl ResGPU {
                 i18n("Highest:"),
                 highest_temperature_string
             ));
-            imp.temperature.graph().push_data_point(*temperature as f64);
+            imp.temperature.graph().push_data_point(*temperature);
 
             usage_percentage_string.push_str(" · ");
             usage_percentage_string.push_str(&temperature_string);
